Webinar hero background version & mobile/tablet layout spacing
