+++ website/src/components/webinars/WebinarLayout.module.scss
@@ -1,0 +1,4 @@
+.hero {
+    background: url(/bg-black-grid.svg) no-repeat;
+    background-size: cover;
+}
